+++ Source/FunicularSwitch.Generators/Transformer/InvokeMethod.cs
@@ -1,0 +1,3 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal delegate string InvokeMethod(IReadOnlyList<string> typeParameters, IReadOnlyList<string> parameters);
