+++ Source/FunicularSwitch.Generators/Transformer/MethodBody.cs
@@ -1,0 +1,8 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal readonly record struct MethodBody(string Value)
+{
+    public static implicit operator MethodBody(string value) => new(value);
+
+    public override string ToString() => Value;
+}
