+++ Source/FunicularSwitch.Generators/Transformer/MethodInfo.cs
@@ -1,0 +1,5 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal record MethodInfo(
+    string Name,
+    InvokeMethod Invoke);
