+++ Source/FunicularSwitch.Generators/Transformer/MonadData.cs
@@ -1,0 +1,7 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal record MonadData(
+    Func<string, string> GenericTypeName,
+    MethodInfo ReturnMethod,
+    MethodInfo BindMethod,
+    bool ImplementsMonadInterface = false);
