+++ Source/FunicularSwitch.Generators/Transformer/MonadImplementationGenerationInfo.cs
@@ -1,0 +1,5 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal record MonadImplementationGenerationInfo(
+    Func<string, string> GenericTypeName,
+    MonadData Monad);
