+++ Source/FunicularSwitch.Generators/Transformer/ParameterGenerationInfo.cs
@@ -1,0 +1,6 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal record ParameterGenerationInfo(
+    string Type,
+    string Name,
+    bool IsExtension = false);
