+++ Source/FunicularSwitch.Generators/Transformer/StaticMonadGenerationInfo.cs
@@ -1,0 +1,7 @@
+namespace FunicularSwitch.Generators.Transformer;
+
+internal record StaticMonadGenerationInfo(
+    string TypeName,
+    string AccessModifier,
+    IReadOnlyList<MonadImplementationGenerationInfo> MonadsWithoutImplementation,
+    IReadOnlyList<MethodGenerationInfo> Methods);
